Refactors control settings behavior
Changes the way settings elements are added to the event arguments.

Instead of directly visiting and adding the associated object, the code now checks if the settings key can be saved using the `_settingsKeyInteractionHub`. If it can, the element is added to the event arguments.

This provides more control over which settings elements are included, allowing for more flexible and conditional settings loading.